Feat: modify package.json field and scripts & move Provider to stroies
