+++ src/index.ts
@@ -46,7 +46,6 @@ export { default as TextField } from './Input/TextField';
 export { default as Toast } from './Display/Toast';
 export { default as Tooltip } from './Display/Tooltip';
 export { default as Typography } from './General/Typography';
-export { Uikit as default, UikitConsumer, UikitProvider } from './Provider';
 
 // Employers
 export { default as EmployersBlockquote } from './Display/EmployersBlockquote';
